Show forks count instead of stargazers count when sorting by forks count
